minds: migrate desktop client from inline HTML strings to Jinja2 + Tailwind
Inline triple-quoted HTML in templates.py/templates_auth.py was getting
unwieldy -- almost 1600 lines of Python-interleaved markup with JS, ad-hoc
styles, and Jinja conditionals all tangled together. This lifts all of
that into:

  apps/minds/imbue/minds/desktop_client/templates/   (15 .html files)
  apps/minds/imbue/minds/desktop_client/static/      (JS + Tailwind CSS)

templates.py and templates_auth.py are now thin wrappers:
  render_foo(...) -> _JINJA_ENV.get_template('foo.html').render(...)
Public signatures are unchanged, so tests and callers are unchanged.

Tailwind pipeline: `just minds-tailwind` runs the CLI (via pnpm) to
generate static/tailwind.css from the templates' class references. The
output is gitignored -- it's a build artifact that ships with the
distributable. Content-scan-only (no kitchen-sink safelist); builds in
~1s and yields a ~20KB CSS bundle.

JS extracted from templates into per-page static files:
  static/chrome.js     -- chrome titlebar + iframe + embedded sidebar
  static/sidebar.js    -- standalone Electron sidebar WebContentsView
  static/sharing.js    -- sharing editor (ACL diff; DOM-built, not innerHTML)
  static/creating.js   -- agent-creation SSE log stream
  static/auth.js       -- sign-up / sign-in tab handling + OAuth polling

FastAPI mounts StaticFiles at /_static; templates link to /_static/tailwind.css
and /_static/<page>.js. A startup warning fires if tailwind.css is missing
(run `just minds-tailwind` to build it).

Partials:
  templates/base.html              -- shared <head> + body scaffold
  templates/_macros.html           -- button / card / input / spinner / notice
  templates/_associate.html        -- account-association form used by settings
                                      and sharing-before-associate
  templates/auth/_auth_base.html   -- centered auth-card layout
  templates/auth/_oauth_icons.html -- Google/GitHub SVGs

Tests unchanged in spirit; two of them were checking for implementation
details (display: none on mac, inline EventSource) and got updated to
check the migrated equivalents (hidden class, <script src> tag).
Full apps/minds test suite: 550 passed.
