+++ .gitignore
@@ -303,3 +303,8 @@ apps/minds_workspace_server/package-lock.json
 
 # Generated by _generate-release-dockerfile from Dockerfile + .release.extras.
 /libs/mngr/imbue/mngr/resources/Dockerfile.release
+
+# Generated Tailwind CSS for the minds desktop client. Built via
+# `just minds-tailwind` (which runs `pnpm tailwind` inside apps/minds) and
+# shipped with the distributable. Not checked in so we can regenerate freely.
+/apps/minds/imbue/minds/desktop_client/static/tailwind.css
