+++ apps/minds/imbue/minds/desktop_client/static/_tailwind_input.css
@@ -1,0 +1,83 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+/* Design tokens the templates read via var(...). Tailwind covers colors,
+   spacing, typography, and radii; everything else lives here. */
+:root {
+  /* Chrome + content surface custom tints that we want stable across pages. */
+  --bg-chrome: #18181b;
+  --bg-chrome-elev: #27272a;
+  --bg-chrome-hover: rgba(255, 255, 255, 0.06);
+  --bg-chrome-active: rgba(255, 255, 255, 0.10);
+  --border-chrome: rgba(255, 255, 255, 0.08);
+  --text-chrome: #e4e4e7;
+  --text-chrome-muted: #a1a1aa;
+
+  /* Link / focus ring */
+  --link: #2563eb;
+  --focus-ring: 0 0 0 3px rgba(37, 99, 235, 0.18);
+
+  /* Chrome seam shadow */
+  --shadow-seam: 0 4px 10px -6px rgba(0, 0, 0, 0.35);
+}
+
+/* Workspace identity stripe -- 3px band in the project's accent at the very top. */
+.page-workspace::before {
+  content: '';
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  height: 3px;
+  background: var(--workspace-accent, oklch(65% 0.15 230));
+  z-index: 1000;
+}
+
+/* The 3px spine on landing rows + sidebar rows. */
+.accent-spine::before {
+  content: '';
+  position: absolute;
+  left: 0;
+  top: 0;
+  bottom: 0;
+  width: 3px;
+  background: var(--workspace-accent, oklch(65% 0.15 230));
+}
+
+.sidebar-item {
+  position: relative;
+}
+.sidebar-item::before {
+  content: '';
+  position: absolute;
+  left: 4px;
+  top: 8px;
+  bottom: 8px;
+  width: 3px;
+  border-radius: 2px;
+  background: var(--workspace-accent, oklch(65% 0.15 230));
+  opacity: 0.55;
+  transition: opacity 100ms;
+}
+.sidebar-item:hover::before,
+.sidebar-item.is-current::before {
+  opacity: 1;
+}
+
+/* Deterministic per-workspace swatch block for the titlebar. */
+.accent-swatch {
+  background: var(--workspace-accent, oklch(65% 0.15 230));
+}
+
+.spinner {
+  border: 2px solid #e4e4e7;
+  border-top-color: #18181b;
+  border-radius: 50%;
+  animation: spin 800ms linear infinite;
+}
+@keyframes spin {
+  to {
+    transform: rotate(360deg);
+  }
+}
