+++ apps/minds/imbue/minds/desktop_client/templates/auth/forgot_password.html
@@ -1,0 +1,44 @@
+{% extends 'auth/_auth_base.html' %}
+{% block title %}Forgot password{% endblock %}
+{% block card_content %}
+  <h1 class="text-2xl font-semibold text-zinc-900 mb-2">Reset password</h1>
+  <p class="text-zinc-500 mb-6">Enter your email to receive a reset link</p>
+  <div id="error-msg" class="hidden text-red-600 text-sm mb-3 px-3 py-2 bg-red-50 rounded-md"></div>
+  <div id="success-msg" class="hidden text-emerald-700 text-sm mb-3 px-3 py-2 bg-emerald-50 rounded-md border border-emerald-200"></div>
+  <form id="forgot-form">
+    <div class="mb-4">
+      <label for="email" class="block text-sm font-medium text-zinc-900 mb-1.5">Email</label>
+      <input type="email" id="email" required autocomplete="email"
+             class="w-full px-3 py-2.5 text-sm rounded-lg border border-zinc-200 bg-white text-zinc-900 outline-none transition focus:border-blue-600 focus:ring-2 focus:ring-blue-600/15">
+    </div>
+    <button type="submit" id="submit-btn"
+            class="w-full py-3 rounded-lg bg-zinc-900 text-zinc-50 font-semibold cursor-pointer transition-colors hover:bg-zinc-800 disabled:bg-zinc-400 disabled:cursor-not-allowed">Send reset link</button>
+  </form>
+  <div class="text-center mt-5 text-sm text-zinc-500"><a href="/auth/login" class="text-blue-600 font-medium hover:underline">Back to sign in</a></div>
+{% endblock %}
+{% block scripts %}
+  <script>
+    document.getElementById('forgot-form').addEventListener('submit', async function (e) {
+      e.preventDefault();
+      var btn = document.getElementById('submit-btn');
+      btn.disabled = true;
+      document.getElementById('error-msg').classList.add('hidden');
+      try {
+        var res = await fetch('/auth/api/forgot-password', {
+          method: 'POST',
+          headers: { 'Content-Type': 'application/json' },
+          body: JSON.stringify({ email: document.getElementById('email').value }),
+        });
+        var data = await res.json();
+        var s = document.getElementById('success-msg');
+        s.textContent = data.message;
+        s.classList.remove('hidden');
+      } catch (err) {
+        var el = document.getElementById('error-msg');
+        el.textContent = 'Network error';
+        el.classList.remove('hidden');
+      }
+      btn.disabled = false;
+    });
+  </script>
+{% endblock %}
