+++ apps/minds/imbue/minds/desktop_client/templates/auth/oauth_close.html
@@ -1,0 +1,9 @@
+{% extends 'auth/_auth_base.html' %}
+{% block title %}Signed in{% endblock %}
+{% block card_content %}
+  <div class="text-center">
+    <h1 class="text-2xl font-semibold text-zinc-900 mb-2">Signed in</h1>
+    <p class="text-zinc-500 mb-3">Signed in as <strong>{{ display_name or email }}</strong></p>
+    <p class="text-sm text-zinc-500">You can close this tab and return to the app.</p>
+  </div>
+{% endblock %}
