+++ apps/minds/imbue/minds/desktop_client/templates/landing.html
@@ -1,0 +1,45 @@
+{% extends 'base.html' %}
+{% import '_macros.html' as ui %}
+{% block title %}Projects{% endblock %}
+{% block content %}
+  {% call ui.page_container() %}
+    {% if agent_ids %}
+      <div class="flex items-center justify-between mb-5">
+        <h1 class="text-xl font-semibold text-zinc-900">Projects</h1>
+        {{ ui.btn_link('Create', '/create', variant='primary') }}
+      </div>
+
+      <div class="flex flex-col gap-1.5">
+        {% for agent_id in agent_ids %}
+          <div class="accent-spine relative flex items-center gap-3 bg-white border border-zinc-200 rounded-xl px-4 py-3.5 cursor-pointer overflow-hidden transition hover:border-zinc-300 hover:shadow-sm"
+               style="--workspace-accent: {{ agent_accents.get(agent_id | string, 'oklch(65% 0.15 230)') }};"
+               data-agent-id="{{ agent_id }}"
+               onclick="window.location='/goto/{{ agent_id }}/'">
+            <span class="flex-1 font-medium text-zinc-900 pl-1">{{ agent_names.get(agent_id | string, agent_id) }}</span>
+            <button class="bg-transparent border border-transparent rounded-md cursor-pointer p-1.5 text-zinc-400 hover:bg-zinc-100 hover:text-zinc-600 flex items-center justify-center"
+                    onclick="event.stopPropagation(); window.location='/workspace/{{ agent_id }}/settings'"
+                    title="Settings"
+                    aria-label="Workspace settings">
+              <svg class="w-4 h-4" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
+                <circle cx="12" cy="12" r="3"/>
+                <path d="M19.4 15a1.65 1.65 0 0 0 .33 1.82l.06.06a2 2 0 0 1-2.83 2.83l-.06-.06a1.65 1.65 0 0 0-1.82-.33 1.65 1.65 0 0 0-1 1.51V21a2 2 0 0 1-4 0v-.09A1.65 1.65 0 0 0 9 19.4a1.65 1.65 0 0 0-1.82.33l-.06.06a2 2 0 0 1-2.83-2.83l.06-.06A1.65 1.65 0 0 0 4.68 15a1.65 1.65 0 0 0-1.51-1H3a2 2 0 0 1 0-4h.09A1.65 1.65 0 0 0 4.6 9a1.65 1.65 0 0 0-.33-1.82l-.06-.06a2 2 0 0 1 2.83-2.83l.06.06A1.65 1.65 0 0 0 9 4.68a1.65 1.65 0 0 0 1-1.51V3a2 2 0 0 1 4 0v.09a1.65 1.65 0 0 0 1 1.51 1.65 1.65 0 0 0 1.82-.33l.06-.06a2 2 0 0 1 2.83 2.83l-.06.06A1.65 1.65 0 0 0 19.4 9a1.65 1.65 0 0 0 1.51 1H21a2 2 0 0 1 0 4h-.09a1.65 1.65 0 0 0-1.51 1z"/>
+              </svg>
+            </button>
+          </div>
+        {% endfor %}
+      </div>
+    {% else %}
+      {% if is_discovering %}
+        <div class="flex items-center justify-center min-h-[80vh]">
+          <p class="text-zinc-400 text-center">Discovering agents...</p>
+        </div>
+        <script>setTimeout(function() { location.reload(); }, 2000);</script>
+      {% else %}
+        <div class="text-center py-12">
+          <p class="text-zinc-400 mb-6">No projects yet</p>
+          {{ ui.btn_link('Create', '/create', variant='primary') }}
+        </div>
+      {% endif %}
+    {% endif %}
+  {% endcall %}
+{% endblock %}
