+++ apps/minds/imbue/minds/desktop_client/templates/login.html
@@ -1,0 +1,13 @@
+{% extends 'base.html' %}
+{% block title %}Login - Projects{% endblock %}
+{% block body_class %}bg-zinc-50 text-zinc-900 font-sans antialiased flex items-center justify-center min-h-screen{% endblock %}
+{% block content %}
+  <div class="bg-white border border-zinc-200 rounded-xl shadow-sm p-8 max-w-[420px] w-full m-4">
+    <h1 class="text-xl font-semibold text-zinc-900 leading-tight">Sign in to Minds</h1>
+    <p class="text-xs text-zinc-400 mt-1.5 mb-5">Use the login URL printed in the terminal.</p>
+    <p class="text-zinc-600 leading-relaxed">
+      Each login URL can only be used once. If you've already used yours, restart the server to
+      generate a new one.
+    </p>
+  </div>
+{% endblock %}
