+++ apps/minds/package.json
@@ -8,13 +8,15 @@
   "scripts": {
     "start": "electron .",
     "build": "node scripts/build.js",
-    "dist": "pnpm build && pnpm exec todesktop build"
+    "dist": "pnpm build && pnpm exec todesktop build",
+    "tailwind": "tailwindcss -i imbue/minds/desktop_client/static/_tailwind_input.css -o imbue/minds/desktop_client/static/tailwind.css --minify"
   },
   "dependencies": {
     "@todesktop/runtime": "^1.6.0"
   },
   "devDependencies": {
     "electron": "35.7.5",
-    "@todesktop/cli": "^1.8.0"
+    "@todesktop/cli": "^1.8.0",
+    "tailwindcss": "^3.4.17"
   }
 }
